refactor(compiler): misc i18n refactoring
