+++ packages/compiler/src/i18n/i18n_html_parser.ts
@@ -12,11 +12,7 @@ import {DEFAULT_INTERPOLATION_CONFIG, InterpolationConfig} from '../ml_parser/in
 import {ParseTreeResult} from '../ml_parser/parser';
 import {sha1Digest} from './digest';
 import {mergeTranslations} from './extractor_merger';
-import {Serializer} from './serializers/serializer';
-import {Xliff} from './serializers/xliff';
-import {Xliff2} from './serializers/xliff2';
-import {Xmb} from './serializers/xmb';
-import {Xtb} from './serializers/xtb';
+import {createSerializer} from './serializers/factory';
 import {TranslationBundle} from './translation_bundle';
 
 export class I18NHtmlParser implements HtmlParser {
@@ -55,21 +51,5 @@ export class I18NHtmlParser implements HtmlParser {
 
     return mergeTranslations(
         parseResult.rootNodes, this._translationBundle, interpolationConfig, [], {});
-  }
-}
-
-export function createSerializer(format: string, version: I18nVersion): Serializer {
-  switch (format) {
-    case 'xmb':
-      return new Xmb(version);
-    case 'xtb':
-      return new Xtb(version);
-    case 'xliff2':
-    case 'xlf2':
-      return new Xliff2(version);
-    case 'xliff':
-    case 'xlf':
-    default:
-      return new Xliff(version);
   }
 }
