+++ packages/compiler/src/i18n/index.ts
@@ -7,8 +7,9 @@
  */
 
 export {Extractor, ExtractorHost} from './extractor';
-export {I18NHtmlParser, createSerializer} from './i18n_html_parser';
+export {I18NHtmlParser} from './i18n_html_parser';
 export {MessageBundle} from './message_bundle';
+export {createSerializer} from './serializers/factory';
 export {Serializer} from './serializers/serializer';
 export {Xliff} from './serializers/xliff';
 export {Xliff2} from './serializers/xliff2';
